+++ source/Public/Contact Management/List/Remove-ContactList.ps1
@@ -1,0 +1,31 @@
+function Remove-ContactList {
+    <#
+    .SYNOPSIS
+    Removes a contact list by its ID.
+
+    .DESCRIPTION
+    The Remove-ContactList function deletes a contact list identified by its unique ID. 
+    It sends a DELETE request to the specified URI endpoint.
+
+    .PARAMETER listId
+    The ID of the contact list to delete. This parameter is mandatory.
+
+    .EXAMPLE
+    Remove-ContactList -listId 12345
+    This command deletes the contact list with the ID 12345.
+    #>
+    [CmdletBinding()]
+    param (
+        [Parameter(Mandatory = $true, HelpMessage = "The ID of the contact list to delete")]
+        [int]$listId
+    )
+    $uri = "/contacts/lists/$listId"
+    $method = "DELETE"
+    
+    $Params = @{
+        "URI"    = $uri
+        "Method" = $method
+    }
+    $list = Invoke-BrevoCall @Params
+    return $list
+}