+++ CHANGELOG.md
@@ -6,16 +6,40 @@
 
 ### Added
 
+- function Add-ContactListMember
+- function Confirm-Domain
 - function Connect-Brevo
+- function Get-Account
+- function Get-Contact
 - function Get-ContactAttribute
 - function Get-ContactFolder
 - function Get-ContactList
+- function Get-ContactListMember
+- function Get-ContactSegment
+- function Get-Domain
+- function Get-EmailCampain
+- function Get-Sender
+- function Get-User
+- function Get-UserActivitylog
+- function Get-UserPermission
+- function Import-Contact
 - function Invoke-BrevoCall
 - function New-Contact
 - function New-ContactAttribute
 - function New-ContactFolder
 - function New-ContactList
+- function New-Domain
+- function New-Sender
+- function Remove-Contact
+- function Remove-ContactList
+- function Remove-Domain
+- function Remove-EmailCampain
+- function Remove-Sender
+- function Send-EmailCampain
+- function Send-UserInvitation
+- function Test-Domain
 - function Update-Contact
+- function Update-ContactList
 
 ### Changed